requirements.txt
streamlit
PyPDF2
google-generativeai
nltk
spacy
scikit-learn
numpy
pandas